.ts and .tsx file  are destructured
