+++ src/index.ts
@@ -1,1 +1,3 @@
-export { GapiAuth, useGapiAuthLogin, useGapiAuthLogout, useGapiAuthUser } from './web/google';
+export { GapiAuth } from './web/google';
+export { useGapiAuthLogin, useGapiAuthLogout, useGapiAuthUser } from './web/google/hooks'
+
