+++ src/web/google/index.tsx
@@ -1,13 +1,13 @@
 import React, { createContext, useContext, useEffect, useState } from 'react';
 import loadScript from './load-script';
-// import removeScript from './remove-script';
+import removeScript from './remove-script';
 
-const AuthContext = createContext<typeof gapi.auth2 | null>(null);
+export const AuthContext = createContext<typeof gapi.auth2 | null>(null);
 export function GapiAuth({ clientId, ...other }: { clientId: string }) {
   const [auth2, setAuth2] = useState<typeof gapi.auth2 | null>(null);
 
   useEffect(() => {
-    let mounted = false;
+    let mounted = true;
     loadScript(
       document,
       'script',
@@ -23,7 +23,7 @@ export function GapiAuth({ clientId, ...other }: { clientId: string }) {
           console.log('google auth', GoogleAuth);
           if (!GoogleAuth) {
             window.gapi.auth2.init(params).then(res => {
-              if (!mounted) {
+              if (mounted) {
                 setAuth2(window.gapi.auth2)
               }
             },
@@ -38,8 +38,8 @@ export function GapiAuth({ clientId, ...other }: { clientId: string }) {
       }
     )
     return () => {
-      mounted = true
-      // removeScript(document, 'google-login')
+      mounted = false
+      removeScript(document, 'google-login')
     }
   }, [])
 
@@ -49,38 +49,9 @@ export function GapiAuth({ clientId, ...other }: { clientId: string }) {
 }
 
 
-export function useGapiAuthLogin() {
-  const auth2 = useContext(AuthContext);
 
-  function signIn(e?: any) {
-    if (e) {
-      e.preventDefault() // to prevent submit if used within form
-    }
-    auth2?.getAuthInstance().signIn().then(
-      res => console.log(res),
-      err => console.log(err)
-    )
-  }
-  return signIn as () => void
-}
 
-export function useGapiAuthLogout() {
-  const auth2 = useContext(AuthContext);
-  const signOut = () => {
-    auth2?.getAuthInstance().signOut().then(() => {
-      auth2?.getAuthInstance().disconnect()
-    })
-  }
-  return signOut as () => void
-}
 
-export function useGapiAuthUser() {
-  const auth2 = useContext(AuthContext);
-  const userObject = auth2?.getAuthInstance().currentUser.get();
-
-  if (userObject === null) return null
-  return userObject
-}
