+++ src/web/google/remove-script.ts
@@ -1,4 +1,4 @@
-export default (d: Document, id:string) => {
+export default (d:Document, id:string) => {
   const element = d.getElementById(id)
 
   if (element) {
